change types to stdint types

+++ philo/src/error_handler.c
@@ -12,7 +12,7 @@
 
 #include <philo.h>
 
-int	thrw_err(char *err_msg, char *err_file, int err_line)
+uint16_t	thrw_err(char *err_msg, char *err_file, int err_line)
 {
 	printf("%s:	%s:%d\n", err_msg, err_file, err_line);
 	printf(USAGE_MSG);
